Fix compiler warnings

+++ src/alg_misc.cc
@@ -29,12 +29,8 @@ bool is_alg_quotient(const alg_t<int64_t> &p1, const alg_t<int64_t> &p2,
                      algebraicQ &ratio)  // if p1/p2 is algebraic, return true
                                          // and put it in ratio.
 {
-  //    out_s << endl << "In is_alg_quotient, p1 = " << p1 << " p2 = " << p2 <<
-  //    endl;
-
-  int i;
+  size_t i;
   algebraicI q(p1.coeffs[0].field());
-  //   algebraic<int64_t> zero_F = algebraic<int64_t>(p1.coeffs[0].field());
 
   for (i = 0; i <= Params::nbr_params(); i++) {
     q = p2.coeffs[i];
@@ -49,10 +45,8 @@ bool is_alg_quotient(const alg_t<int64_t> &p1, const alg_t<int64_t> &p2,
 
   algebraicQ num(to_rational(p1.coeffs[i]));
   algebraicQ denom(to_rational(p2.coeffs[i]));
-  //    out_s << "num = " << num << " denom = " << denom << endl;
 
   ratio = num / denom;
-  //  out_s << "Ratio is : " << ratio << flush;
 
   for (i = 0; i <= Params::nbr_params(); i++) {
     algebraicQ denom_i = to_rational(p2.coeffs[i]);
@@ -62,7 +56,6 @@ bool is_alg_quotient(const alg_t<int64_t> &p1, const alg_t<int64_t> &p2,
       return (false);
     }
   }
-  //    out_s << "ratio = " << endl ;
   return (true);
 }
 }  // namespace polygon
